backend/requirements.txt
Flask==2.0.3
flask-cors==3.0.10
# sentence-transformers==2.1.0
openai==0.28.0
qdrant-client==0.9.0
google-cloud-tasks==2.7.1
google-generativeai==0.1.0
python-dotenv==0.19.2
gunicorn==20.1.0
Werkzeug==2.0.3